+++ plugins/woocommerce-admin/includes/class-wc-admin-api-init.php
@@ -87,6 +87,9 @@ class WC_Admin_Api_Init {
 	public function rest_api_init() {
 		require_once dirname( __FILE__ ) . '/api/class-wc-admin-rest-admin-notes-controller.php';
 		require_once dirname( __FILE__ ) . '/api/class-wc-admin-rest-customers-controller.php';
+		require_once dirname( __FILE__ ) . '/api/class-wc-admin-rest-data-controller.php';
+		require_once dirname( __FILE__ ) . '/api/class-wc-admin-rest-data-download-ips-controller.php';
+		require_once dirname( __FILE__ ) . '/api/class-wc-admin-rest-orders-controller.php';
 		require_once dirname( __FILE__ ) . '/api/class-wc-admin-rest-products-controller.php';
 		require_once dirname( __FILE__ ) . '/api/class-wc-admin-rest-product-reviews-controller.php';
 		require_once dirname( __FILE__ ) . '/api/class-wc-admin-rest-reports-controller.php';
@@ -111,6 +114,9 @@ class WC_Admin_Api_Init {
 		$controllers = array(
 			'WC_Admin_REST_Admin_Notes_Controller',
 			'WC_Admin_REST_Customers_Controller',
+			'WC_Admin_REST_Data_Controller',
+			'WC_Admin_REST_Data_Download_Ips_Controller',
+			'WC_Admin_REST_Orders_Controller',
 			'WC_Admin_REST_Products_Controller',
 			'WC_Admin_REST_Product_Reviews_Controller',
 			'WC_Admin_REST_Reports_Controller',
@@ -170,17 +176,6 @@ class WC_Admin_Api_Init {
 			$endpoints['/wc/v3/reports'][0] = $endpoints['/wc/v3/reports'][1];
 		}
 
-		// Override /wc/v3/products.
-		if ( isset( $endpoints['/wc/v3/products'] )
-			&& isset( $endpoints['/wc/v3/products'][3] )
-			&& isset( $endpoints['/wc/v3/products'][2] )
-			&& $endpoints['/wc/v3/products'][2]['callback'][0] instanceof WC_Admin_REST_Products_Controller
-			&& $endpoints['/wc/v3/products'][3]['callback'][0] instanceof WC_Admin_REST_Products_Controller
-		) {
-			$endpoints['/wc/v3/products'][0] = $endpoints['/wc/v3/products'][2];
-			$endpoints['/wc/v3/products'][1] = $endpoints['/wc/v3/products'][3];
-		}
-
 		// Override /wc/v3/customers.
 		if ( isset( $endpoints['/wc/v3/customers'] )
 			&& isset( $endpoints['/wc/v3/customers'][3] )
@@ -190,6 +185,50 @@ class WC_Admin_Api_Init {
 		) {
 			$endpoints['/wc/v3/customers'][0] = $endpoints['/wc/v3/customers'][2];
 			$endpoints['/wc/v3/customers'][1] = $endpoints['/wc/v3/customers'][3];
+		}
+
+		// Override /wc/v3/orders/$id.
+		if ( isset( $endpoints['/wc/v3/orders/(?P<id>[\d]+)'] )
+			&& isset( $endpoints['/wc/v3/orders/(?P<id>[\d]+)'][5] )
+			&& isset( $endpoints['/wc/v3/orders/(?P<id>[\d]+)'][4] )
+			&& isset( $endpoints['/wc/v3/orders/(?P<id>[\d]+)'][3] )
+			&& $endpoints['/wc/v3/orders/(?P<id>[\d]+)'][3]['callback'][0] instanceof WC_Admin_REST_Orders_Controller
+			&& $endpoints['/wc/v3/orders/(?P<id>[\d]+)'][4]['callback'][0] instanceof WC_Admin_REST_Orders_Controller
+			&& $endpoints['/wc/v3/orders/(?P<id>[\d]+)'][5]['callback'][0] instanceof WC_Admin_REST_Orders_Controller
+		) {
+			$endpoints['/wc/v3/orders/(?P<id>[\d]+)'][0] = $endpoints['/wc/v3/orders/(?P<id>[\d]+)'][3];
+			$endpoints['/wc/v3/orders/(?P<id>[\d]+)'][1] = $endpoints['/wc/v3/orders/(?P<id>[\d]+)'][4];
+			$endpoints['/wc/v3/orders/(?P<id>[\d]+)'][2] = $endpoints['/wc/v3/orders/(?P<id>[\d]+)'][5];
+		}
+
+		// Override /wc/v3orders.
+		if ( isset( $endpoints['/wc/v3/orders'] )
+			&& isset( $endpoints['/wc/v3/orders'][3] )
+			&& isset( $endpoints['/wc/v3/orders'][2] )
+			&& $endpoints['/wc/v3/orders'][2]['callback'][0] instanceof WC_Admin_REST_Orders_Controller
+			&& $endpoints['/wc/v3/orders'][3]['callback'][0] instanceof WC_Admin_REST_Orders_Controller
+		) {
+			$endpoints['/wc/v3/orders'][0] = $endpoints['/wc/v3/orders'][2];
+			$endpoints['/wc/v3/orders'][1] = $endpoints['/wc/v3/orders'][3];
+		}
+
+		// Override /wc/v3/data.
+		if ( isset( $endpoints['/wc/v3/data'] )
+			&& isset( $endpoints['/wc/v3/data'][1] )
+			&& $endpoints['/wc/v3/data'][1]['callback'][0] instanceof WC_Admin_REST_Data_Controller
+		) {
+			$endpoints['/wc/v3/data'][0] = $endpoints['/wc/v3/data'][1];
+		}
+
+		// Override /wc/v3/products.
+		if ( isset( $endpoints['/wc/v3/products'] )
+			&& isset( $endpoints['/wc/v3/products'][3] )
+			&& isset( $endpoints['/wc/v3/products'][2] )
+			&& $endpoints['/wc/v3/products'][2]['callback'][0] instanceof WC_Admin_REST_Products_Controller
+			&& $endpoints['/wc/v3/products'][3]['callback'][0] instanceof WC_Admin_REST_Products_Controller
+		) {
+			$endpoints['/wc/v3/products'][0] = $endpoints['/wc/v3/products'][2];
+			$endpoints['/wc/v3/products'][1] = $endpoints['/wc/v3/products'][3];
 		}
 
 		// Override /wc/v3/products/$id.