+++ plugins/woocommerce-admin/packages/components/src/search/autocompleters/index.js
@@ -5,7 +5,9 @@
 export { default as countries } from './countries';
 export { default as coupons } from './coupons';
 export { default as customers } from './customers';
+export { default as downloadIps } from './download-ips';
 export { default as emails } from './emails';
+export { default as orders } from './orders';
 export { default as product } from './product';
 export { default as productCategory } from './product-cat';
 export { default as taxes } from './taxes';